refactor(react): replace `useStore` with multi-store `useStore`
