+++ packages/@livestore/react/src/experimental/multi-store/mod.ts
@@ -1,5 +1,4 @@
+export * from '../../StoreRegistryContext.tsx'
 export * from './StoreRegistry.ts'
-export * from './StoreRegistryContext.tsx'
 export * from './storeOptions.ts'
 export * from './types.ts'
-export * from './useStore.ts'
