+++ packages/@livestore/react/src/useStore.test.tsx
@@ -5,11 +5,11 @@ import { shouldNeverHappen } from '@livestore/utils'
 import { act, type RenderHookResult, type RenderResult, render, renderHook, waitFor } from '@testing-library/react'
 import * as React from 'react'
 import { describe, expect, it } from 'vitest'
-import { schema } from '../../__tests__/fixture.tsx'
-import { StoreRegistry } from './StoreRegistry.ts'
+import { schema } from './__tests__/fixture.tsx'
+import { StoreRegistry } from './experimental/multi-store/StoreRegistry.ts'
+import { storeOptions } from './experimental/multi-store/storeOptions.ts'
+import type { CachedStoreOptions } from './experimental/multi-store/types.ts'
 import { StoreRegistryProvider } from './StoreRegistryContext.tsx'
-import { storeOptions } from './storeOptions.ts'
-import type { CachedStoreOptions } from './types.ts'
 import { useStore } from './useStore.ts'
 
 describe('experimental useStore', () => {
